NEW - bug 149606: non-Java resource fails to show when source folder is not interesting https://bugs.eclipse.org/bugs/show_bug.cgi?id=149606

+++ sandbox/org.eclipse.mylyn.monitor.reports/src/org/eclipse/mylyn/internal/monitor/reports/collectors/ViewUsageCollector.java
@@ -60,7 +60,7 @@ public class ViewUsageCollector implements IUsageCollector {
 	public List<String> getSummary(int userId) {
 		Map<String, Integer> normalViewSelections = usersNormalViewSelections.get(userId);
 
-		float numSelections = (float) usersNumSelections.get(userId);
+		float numSelections = usersNumSelections.get(userId);
 
 		List<String> summaries = new ArrayList<String>();
 		List<String> viewUsage = new ArrayList<String>();
